windows compatibility

+++ ething/webserver/WebServer.py
@@ -1,9 +1,14 @@
 # coding: utf-8
-from multiprocessing import Process
 import os
 import signal
 
-    
+is_thread = os.name == "nt"
+
+if is_thread:
+    from threading import Thread
+else:
+    from multiprocessing import Process
+
 
 class WebServer(object):
     
@@ -16,12 +21,13 @@ class WebServer(object):
     @property
     def is_started (self):
         return self.p is not None and self.p.is_alive()
-    
+
     def __proc(self):
         
         # remove any signal handlers set by the parents
-        signal.signal(signal.SIGINT, signal.SIG_IGN)
-        signal.signal(signal.SIGTERM, signal.SIG_DFL)
+        if not is_thread:
+            signal.signal(signal.SIGINT, signal.SIG_IGN)
+            signal.signal(signal.SIGTERM, signal.SIG_DFL)
         
         self.core.log.info("webserver process started, pid = %d" % os.getpid() )
         
@@ -37,18 +43,23 @@ class WebServer(object):
         self.core.log.info("webserver process ended")
     
     def start(self):
+        if self.is_started:
+            raise Exception('the webserver is already running !')
+
         #run webserver in another process
-        self.p = Process(target=self.__proc, args=())
+        if is_thread:
+            self.p = Thread(target=self.__proc, args=())
+        else:
+            self.p = Process(target=self.__proc, args=())
+
         self.p.daemon = True
         self.p.start()
     
     def stop(self):
         if self.is_started:
-            self.core.log.info("stopping webserver")
-            self.p.terminate()
-            self.p.join()
-            self.core.log.info("webserver stopped")
-            self.p = None
-        
-        
-        
+            if not is_thread:
+                self.core.log.info("stopping webserver")
+                self.p.terminate()
+                self.p.join()
+                self.core.log.info("webserver stopped")
+                self.p = None
